Syncing from upstream odoo/odoo (master)

+++ addons/stock/models/stock_picking.py
@@ -406,8 +406,7 @@ class Picking(models.Model):
         check_company=True, required=True)
     move_ids = fields.One2many('stock.move', 'picking_id', string="Stock Moves", copy=True)
     move_ids_without_package = fields.One2many(
-        'stock.move', 'picking_id', string="Stock moves not in package", compute='_compute_move_without_package',
-        inverse='_set_move_without_package', compute_sudo=True)
+        'stock.move', 'picking_id', string="Stock moves not in package", domain=['|', ('package_level_id', '=', False), ('picking_type_entire_packs', '=', False)])
     has_scrap_move = fields.Boolean(
         'Has Scrap Moves', compute='_has_scrap_move')
     picking_type_id = fields.Many2one(
@@ -676,12 +675,15 @@ class Picking(models.Model):
         for picking in self:
             picking.has_packages = bool(cnt_by_picking.get(picking.id, False))
 
-    @api.depends('state')
+    @api.depends('state', 'move_ids.product_uom_qty', 'picking_type_code')
     def _compute_show_check_availability(self):
         """ According to `picking.show_check_availability`, the "check availability" button will be
         displayed in the form view of a picking.
         """
         for picking in self:
+            if picking.immediate_transfer and picking.picking_type_code == 'incoming':
+                picking.show_check_availability = False
+                continue
             if picking.state not in ('confirmed', 'waiting', 'assigned'):
                 picking.show_check_availability = False
                 continue
@@ -864,7 +866,7 @@ class Picking(models.Model):
             after_vals['partner_id'] = vals['partner_id']
         if after_vals:
             self.move_ids.filtered(lambda move: not move.scrapped).write(after_vals)
-        if vals.get('move_ids'):
+        if vals.get('move_ids') or vals.get('move_ids_without_package'):
             self._autoconfirm_picking()
 
         return res
@@ -930,9 +932,12 @@ class Picking(models.Model):
         picking_to_reset = self.filtered(lambda p: p.state not in ('done', 'cancel'))
         picking_to_reset.do_unreserve()
         picking_to_reset.immediate_transfer = False
-        picking_to_reset.move_ids.quantity_done = 0
-        picking_to_reset.move_ids.move_line_ids.unlink()
-        picking_to_reset.move_ids.state = 'draft'
+        if picking_to_reset.move_ids:
+            picking_to_reset.move_ids.quantity_done = 0
+            picking_to_reset.move_ids.move_line_ids.unlink()
+            picking_to_reset.move_ids.state = 'draft'
+        else:
+            picking_to_reset.state = 'draft'
 
     def _action_done(self):
         """Call `_action_done` on the `stock.move` of the `stock.picking` in `self`.
@@ -970,35 +975,6 @@ class Picking(models.Model):
                 email_layout_xmlid='mail.mail_notification_light',
                 subtype_id=subtype_id,
             )
-
-    @api.depends('state', 'move_ids', 'move_ids.state', 'move_ids.package_level_id', 'move_ids.move_line_ids.package_level_id')
-    def _compute_move_without_package(self):
-        for picking in self:
-            picking.move_ids_without_package = picking._get_move_ids_without_package()
-
-    def _set_move_without_package(self):
-        new_mwp = self[0].move_ids_without_package
-        for picking in self:
-            old_mwp = picking._get_move_ids_without_package()
-            picking.move_ids = (picking.move_ids - old_mwp) | new_mwp
-            moves_to_unlink = old_mwp - new_mwp
-            if moves_to_unlink:
-                moves_to_unlink.unlink()
-
-    def _get_move_ids_without_package(self):
-        self.ensure_one()
-        move_ids_without_package = self.env['stock.move']
-        if not self.picking_type_entire_packs:
-            move_ids_without_package = self.move_ids
-        else:
-            for move in self.move_ids:
-                if not move.package_level_id:
-                    if move.state == 'assigned' and move.picking_id or move.state == 'done':
-                        if any(not ml.package_level_id for ml in move.move_line_ids):
-                            move_ids_without_package |= move
-                    else:
-                        move_ids_without_package |= move
-        return move_ids_without_package.filtered(lambda move: not move.scrap_id)
 
     def _check_move_lines_map_quant_package(self, package):
         return package._check_move_lines_map_quant(self.move_line_ids.filtered(lambda ml: ml.package_id == package), 'reserved_qty')
@@ -1048,6 +1024,10 @@ class Picking(models.Model):
                         })
                         for pl in package_level_ids:
                             pl.location_dest_id = self._get_entire_pack_location_dest(pl.move_line_ids) or picking.location_dest_id.id
+                        for move in move_lines_to_pack.move_id:
+                            if all(line.package_level_id for line in move.move_line_ids) \
+                                    and len(move.move_line_ids.package_level_id) == 1:
+                                move.package_level_id = move.move_line_ids.package_level_id
 
     def _get_lot_move_lines_for_sanity_check(self, none_done_picking_ids, separate_pickings=True):
         """ Get all move_lines with tracked products that need to be checked over in the sanity check.
@@ -1093,7 +1073,7 @@ class Picking(models.Model):
                 no_quantities_done_ids.add(picking.id)
             if all(float_is_zero(move_line.reserved_qty, precision_rounding=move_line.product_uom_id.rounding) for move_line in picking.move_line_ids):
                 no_reserved_quantities_ids.add(picking.id)
-        pickings_without_quantities = self.filtered(lambda p: p.id in no_quantities_done_ids and p.id in no_reserved_quantities_ids)
+        pickings_without_quantities = self.filtered(lambda p: p.id in no_quantities_done_ids and p.id in no_reserved_quantities_ids and not p.immediate_transfer)
 
         pickings_using_lots = self.filtered(lambda p: p.picking_type_id.use_create_lots or p.picking_type_id.use_existing_lots)
         if pickings_using_lots:
@@ -1190,6 +1170,18 @@ class Picking(models.Model):
             pickings_to_backorder = self._check_backorder()
             if pickings_to_backorder:
                 return pickings_to_backorder._action_generate_backorder_wizard(show_transfers=self._should_show_transfers())
+
+        message = ''
+        for move in self.move_ids:
+            if float_compare(move.quantity_done, move.product_uom_qty, precision_rounding=move.product_uom.rounding) > 0 \
+                    and not float_is_zero(move.product_uom_qty, precision_rounding=move.product_uom.rounding) \
+                    and move.from_immediate_transfer:
+                message += _('%(name)s has a quantity of %(initial_qty)s %(uom)s done, but %(done_qty)s %(uom)s reserved\n', name=move.product_id.display_name, initial_qty=move.product_uom_qty, done_qty=move.quantity_done, uom=move.product_uom.name)
+            if move.from_immediate_transfer and float_compare(move.product_qty, move.quantity_done, precision_rounding=move.product_uom.rounding) > 0:
+                move.quantity_done = move.product_qty
+        if message:
+            message = _('The product(s)\n') + message + _('Update quantities of those products.')
+            raise UserError(message)
         return True
 
     def _should_show_transfers(self):
@@ -1265,6 +1257,8 @@ class Picking(models.Model):
         immediate_pickings = self.browse()
         precision_digits = self.env['decimal.precision'].precision_get('Product Unit of Measure')
         for picking in self:
+            if picking.immediate_transfer:
+                continue
             if all(float_is_zero(move_line.qty_done, precision_digits=precision_digits) for move_line in picking.move_line_ids.filtered(lambda m: m.state not in ('done', 'cancel'))):
                 immediate_pickings |= picking
         return immediate_pickings
@@ -1555,6 +1549,10 @@ class Picking(models.Model):
     def action_put_in_pack(self):
         self.ensure_one()
         if self.state not in ('done', 'cancel'):
+            if self.immediate_transfer:
+                for move in self.move_ids:
+                    if float_is_zero(move.quantity_done, precision_rounding=move.product_uom.rounding):
+                        move.quantity_done = move.product_qty
             picking_move_lines = self.move_line_ids
             if (
                 not self.picking_type_id.show_reserved
